Reorganize ESL tests to follow Ruby conventions
Move test files to match lib directory structure:
- test/unit/switest2/esl_*.rb → test/unit/switest2/esl/*_test.rb
