+++ test/unit/switest2/esl/client_test.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative "../../switest2_test_helper"
+require_relative "../../../switest2_test_helper"
 
 class Switest2::ESL::ClientTest < Minitest::Test
   def setup
