+++ test/unit/switest2/esl/escaper_test.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative "../../switest2_test_helper"
+require_relative "../../../switest2_test_helper"
 
 class Switest2::ESL::EscaperTest < Minitest::Test
   Escaper = Switest2::ESL::Escaper
